Server render park info shell and alert list
