+++ src/app/parks/[park]/components/favorite-checkbox.tsx
