+++ src/app/parks/[park]/components/suspense-image.tsx
