+++ src/app/people/[park]/[person]/components/person-info.tsx
